backend/requirements.txt
# Django & API
Django>=5.0,<6
djangorestframework>=3.15,<4
django-cors-headers>=4.3,<5

# Telegram (async v20+)
python-telegram-bot>=21.0,<22

# DB (PostgreSQL — раскомментируйте при переходе с SQLite)
# psycopg[binary]>=3.1,<4

python-dotenv>=1.0,<2
